// backend/sanity.types.ts
export type Category = {
  _id: string
  _type: 'category'
  _createdAt: string
  _updatedAt: string
  _rev: string
  Id?: string
  name?: string
  slug?: Slug
  parentCategory?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'category'
  }
}
// ...
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol